Created new BaseListRowRender that woudl exented the new ViewHolderBasedRenderer
